# g/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/create-deposit": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Create a deposit transaction",
                "operationId": "4f6dd0cec51b1e73d93fa059c26ffda1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "destination",
                                    "comptId",
                                    "message"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Amount for deposit",
                                        "type": "number"
                                    },
                                    "destination": {
                                        "description": "Recipient's mobile number",
                                        "type": "string"
                                    },
                                    "comptId": {
                                        "description": "Account ID",
                                        "type": "integer"
                                    },
                                    "message": {
                                        "description": "Transaction message",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deposit transaction created successfully"
                    },
                    "400": {
                        "description": "Bad request, missing or invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/get-user-transaction-history": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get user's transaction history",
                "operationId": "e04849fab6f5dd8ccc9e1dc74580aecd",
                "responses": {
                    "200": {
                        "description": "User's transaction history retrieved successfully"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/all-operator-mobiles": {
            "get": {
                "tags": [
                    "Operator Mobile"
                ],
                "summary": "Liste de tous les opérateurs mobiles",
                "operationId": "1d65283103e1977ac27e1440bc7f46d5",
                "responses": {
                    "200": {
                        "description": "Liste des opérateurs mobiles récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-operator-mobile": {
            "post": {
                "tags": [
                    "Operator Mobile"
                ],
                "summary": "Ajouter un nouvel opérateur mobile",
                "operationId": "d2b6b0d7d9d96f4c0abaac457de74ab6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label": {
                                        "description": "Label de l'opérateur mobile",
                                        "type": "string"
                                    },
                                    "logo_url": {
                                        "description": "URL du logo de l'opérateur mobile",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opérateur mobile enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message de succès",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte, des erreurs de validation sont présentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "description": "Liste des erreurs de validation",
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}